C128 TRSE Files
crazyspence.ras - a port of my first C64 program in TRSE from 2020, pretty buggy. WaitForVerticalBlank() doesn't seem to work so I hacked around it
baaars.ras - a port of my NTSC Raster bar C64 file. This was made because the TRSE raster bar demo ran like junk on NTSC so I tweaked it to the point of it even having MORE raster bars. on C128 it can only handle 7 instead of 8, 8 causes a CPU JAM for some reason
tutorial_6_dycp.ras - this tutorial did not work on NTSC C128's at all so I tweaked it into working. Fixed it not running at all, the image blinking, Jibberish text when it finished Still has some artifacts in the text